+++ public/locales/en/common.json
@@ -1024,7 +1024,7 @@
         "bcharge":"Battery Charge",
         "timeleft":"Time Left"
     },
-    "hoarder": {
+    "karakeep": {
         "bookmarks": "Bookmarks",
         "favorites": "Favorites",
         "archived": "Archived",
