+++ src/utils/proxy/handlers/credentialed.js
@@ -41,7 +41,7 @@ export default async function credentialedProxyHandler(req, res, map) {
           "cloudflared",
           "ghostfolio",
           "headscale",
-          "hoarder",
+          "karakeep",
           "linkwarden",
           "mealie",
           "netalertx",
