+++ src/widgets/components.js
@@ -47,7 +47,8 @@ const components = {
   grafana: dynamic(() => import("./grafana/component")),
   hdhomerun: dynamic(() => import("./hdhomerun/component")),
   headscale: dynamic(() => import("./headscale/component")),
-  hoarder: dynamic(() => import("./hoarder/component")),
+  hoarder: dynamic(() => import("./karakeep/component")),
+  karakeep: dynamic(() => import("./karakeep/component")),
   peanut: dynamic(() => import("./peanut/component")),
   homeassistant: dynamic(() => import("./homeassistant/component")),
   homebox: dynamic(() => import("./homebox/component")),
